+++ test/comparison.js
@@ -70,6 +70,74 @@ describe("#eq()", () => {
   });
 });
 
+describe("#ne()", () => {
+  it("returns true if two instances are not equivalent", () => {
+    const cases = [
+      [0n, 1n, 0n, 100n],
+      [10n, 10n, 200n, 200n],
+      [12n, 24n, 720n, 1440n],
+      [43n, 73n, 516n, 876n],
+    ];
+    for (const [xnum, xden, ynum, yden] of cases) {
+      assert(!new BigAmount(xnum, xden).ne(new BigAmount(ynum, yden)));
+      assert(!new BigAmount(-xnum, xden).ne(new BigAmount(-ynum, yden)));
+      assert(!new BigAmount(xnum, -xden).ne(new BigAmount(ynum, -yden)));
+      assert(!new BigAmount(-xnum, -xden).ne(new BigAmount(-ynum, -yden)));
+
+      assert(!new BigAmount(ynum, yden).ne(new BigAmount(xnum, xden)));
+      assert(!new BigAmount(-ynum, yden).ne(new BigAmount(-xnum, xden)));
+      assert(!new BigAmount(ynum, -yden).ne(new BigAmount(xnum, -xden)));
+      assert(!new BigAmount(-ynum, -yden).ne(new BigAmount(-xnum, -xden)));
+    }
+  });
+
+  it("returns true if two instances are not equivalent", () => {
+    const cases = [
+      [0n, 1n, 1n, 100n],
+      [10n, 10n, 201n, 200n],
+      [12n, 24n, 721n, 1440n],
+      [43n, 73n, 517n, 876n],
+    ];
+    for (const [xnum, xden, ynum, yden] of cases) {
+      assert(new BigAmount(xnum, xden).ne(new BigAmount(ynum, yden)));
+      assert(new BigAmount(-xnum, xden).ne(new BigAmount(-ynum, yden)));
+      assert(new BigAmount(xnum, -xden).ne(new BigAmount(ynum, -yden)));
+      assert(new BigAmount(-xnum, -xden).ne(new BigAmount(-ynum, -yden)));
+
+      assert(new BigAmount(ynum, yden).ne(new BigAmount(xnum, xden)));
+      assert(new BigAmount(-ynum, yden).ne(new BigAmount(-xnum, xden)));
+      assert(new BigAmount(ynum, -yden).ne(new BigAmount(xnum, -xden)));
+      assert(new BigAmount(-ynum, -yden).ne(new BigAmount(-xnum, -xden)));
+    }
+  });
+
+  it("returns true if reduced forms are not the same; otherwise, return false", () => {
+    const cases = reduced.map(([num, den, simplestNum, simplestDen]) => [
+      `${simplestNum}/${simplestDen}`,
+      new BigAmount(num, den),
+      new BigAmount(-num, den),
+      new BigAmount(num, -den),
+      new BigAmount(-num, -den),
+    ]);
+
+    for (const x of cases) {
+      for (const y of cases) {
+        if (x[0] === y[0]) {
+          assert(!x[1].ne(y[1]));
+          assert(!x[2].ne(y[2]));
+          assert(!x[3].ne(y[3]));
+          assert(!x[4].ne(y[4]));
+        } else {
+          assert(x[1].ne(y[1]));
+          assert(x[2].ne(y[2]));
+          assert(x[3].ne(y[3]));
+          assert(x[4].ne(y[4]));
+        }
+      }
+    }
+  });
+});
+
 const runOnCmpCases = (test) => {
   const cases = [
     // {{{
@@ -111,10 +179,7 @@ const runOnCmpCases = (test) => {
 describe("cmp()", () => {
   it("compares manually prepared cases properly", () => {
     runOnCmpCases((xn, xd, yn, yd, cmp) => {
-      assert.strictEqual(
-        BigAmount.cmp(new BigAmount(xn, xd), new BigAmount(yn, yd)),
-        cmp
-      );
+      assert.strictEqual(new BigAmount(xn, xd).cmp(new BigAmount(yn, yd)), cmp);
     });
   });
 });